mac/arm64: Add ffmpeg stuff for arm64.
Change opus handling since the current approach didn't work when
cross-building from an intel mac. It sounds like the intent is to
suppress undefined symbol warnings, so I went for a more direct
approach for that, which seems to work. (It requires buildling
with a flat namespace instead of two-level namespaces, but that seems
ok.)

Bug: chromium:1098899
Change-Id: I0b256497dfb2bdd8ca0607d9cb266625e2e3b87e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/third_party/ffmpeg/+/2270709
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>