[mov] See if mfra makes up the difference for an incomplete sidx.
A few popular sites have started generating MP4 files which have a
sidx plus an mfra. The sidx accounts for all size except the mfra,
so the old code did not mark the fragment index as complete.

Instead we can just check if there's an mfra and if its size makes
up the difference we can mark the index as complete.

Bug: 1107130
Change-Id: Ib84a784704a8ad9940893a7ffcf6812d9f9afa2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/third_party/ffmpeg/+/2354154
Reviewed-by: Dan Sanders <sandersd@chromium.org>